% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_loa.R
\name{compare_loa}
\alias{compare_loa}
\title{Title}
\usage{
compare_loa(df, ..., reference, IC_level = 0.95, x_text = NULL,
  y_text = NULL, strip_text = NULL, graph = TRUE)
}
\arguments{
\item{df}{a tibble or a dataframe}

\item{...}{unquoted names of variable to compare}

\item{reference}{unquoted name of reference variable to compare with}

\item{IC_level}{a number. Value of the confidence intervals level}

\item{x_text}{graphic x axis label. If NULL, nothing is displayed}

\item{y_text}{graphic y axis label. If NULL, name of reference variable is displayed}

\item{strip_text}{graphic strip text. IF NULL, names of variables in ... arg are displayed}

\item{graph}{boolean. Should the graphic be plotted ?}
}
\value{
a list with three elements
\item{data}{the data with call data and graphic quantities}
\item{loa}{statistics about limits of agreament}
\item{graphic}{the ggplot graphic}
}
\description{
Title
}
\examples{
#find an example
}
