% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_predictions.R
\name{compare_prediction}
\alias{compare_prediction}
\title{Title}
\usage{
compare_prediction(df, ..., response, success, level = NULL, pal,
  text_class = NULL, graph = TRUE)
}
\arguments{
\item{df}{a tibble or a data frame}

\item{...}{the columns unquoted names with the methods to compare}

\item{response}{the response variable}

\item{success}{a character value defining how the success event of the response variable is coded}

\item{level}{a named vector defining how to code the events of response variable. See example below}

\item{pal}{a vector of colors of length the quantity of methods to compare}

\item{text_class}{a character value. The title response varibale class legend}

\item{graph}{boolean. Should the graphics be plotted ?}
}
\value{
a list with four elements
\item{models}{the different models fitted with some statistics}
\item{roc}{statistics about ROC}
\item{graph_pred}{the  graphic of predictions}
\item{graph_roc}{the graphic of ROC}
}
\description{
Title
}
\examples{
#Find an example
}
