% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rr.R
\name{compare_rr}
\alias{compare_rr}
\title{Function to compare reproducility and repeatability of different measurement methods through the calcul of Intraclass Correlation Coefficient (ICC)}
\usage{
compare_rr(df, ..., subjects, observers, repetitions, which_repetition,
  which_observer, graph = TRUE)
}
\arguments{
\item{df}{a tibble}

\item{...}{unquoted names of column to compare}

\item{subjects}{unquoted name of subject column}

\item{observers}{unquoted name of observers/raters column}

\item{repetitions}{unquoted name of repetitions column}

\item{which_repetition}{a character string. The code of repetitions to use for reproducibility.}

\item{which_observer}{a character string. The code of observers/raters to use for repeatability.}

\item{graph}{boolean. Should the graphics be plotted ?}
}
\value{
a list with two element
\item{Reproducibility}{a list with two element : the tibble and the ggplot graphic of results}
\item{Repeatability}{a list with two element : the tibble and the ggplot graphic of results}
}
\description{
Function to compare reproducility and repeatability of different measurement methods through the calcul of Intraclass Correlation Coefficient (ICC)
}
\examples{
#find example
}
