% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_values.R
\name{compare_values}
\alias{compare_values}
\title{Title}
\usage{
compare_values(df, ..., reference, x_text = NULL, y_text = NULL,
  strip_text = NULL, graph = TRUE)
}
\arguments{
\item{df}{a tibble or a dataframe}

\item{...}{unquoted names of variable to compare}

\item{reference}{unquoted name of reference variable to compare with}

\item{x_text}{graphic x axis label. If NULL, nothing is displayed}

\item{y_text}{graphic y axis label. If NULL, name of reference variable is displayed}

\item{strip_text}{graphic strip text. IF NULL, names of variables in ... arg are displayed}

\item{graph}{boolean. Does the graph should be plotted ?}
}
\value{
a list with two element
\item{models}{a tibble with models estimatoin statistics}
\item{graphic}{the ggplot graphic}
}
\description{
Title
}
\examples{
data(iris)
res <- compare_values(df = iris, Sepal.Length, Petal.Width, reference = Petal.Length, graph = FALSE)
res$models
plot(res$graphic)
}
