% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_confMat}
\alias{get_confMat}
\title{Function to get confusion matrix from a model}
\usage{
get_confMat(model)
}
\arguments{
\item{model}{a model from the train_mod function}
}
\value{
a list. see ?caret::confusionMatrix
}
\description{
Function to get confusion matrix from a model
}
\examples{
#find an example
}
