% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_icc}
\alias{get_icc}
\title{Function to get icc results}
\usage{
get_icc(dd, model, type, unit)
}
\arguments{
\item{dd}{a tibble}

\item{model}{character, type of model. See ?irr::icc}

\item{type}{character, type of assessment. See ?irr::icc}

\item{unit}{character, type of unit. See ?irr::icc}
}
\value{
a tibble with ICC estimate and 95% IC
}
\description{
Function to get icc results
}
\examples{
#find examples
}
