% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_prediction}
\alias{get_prediction}
\title{Function to get predictions according to a fitted model}
\usage{
get_prediction(model, newdata, type)
}
\arguments{
\item{model}{a model from the train_mod function}

\item{newdata}{the dtibble/ataframe with predictors}

\item{type}{"raw" or "prob". see ?predict.train}
}
\value{
a tibble with the data and the corresponding predictions
}
\description{
Function to get predictions according to a fitted model
}
\examples{
#find an example
}
