% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_roc}
\alias{get_roc}
\title{Function to get the ROV statistics of a model}
\usage{
get_roc(model, success)
}
\arguments{
\item{model}{a model from the train_mod function}

\item{success}{character value defining how the success is coded}
}
\value{
a list of class "roc". See ?pROC::roc
}
\description{
Function to get the ROV statistics of a model
}
\examples{
#find an example
}
