% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_senspe}
\alias{get_senspe}
\title{Function to get sensitivities ad specificities from a ROC statistics}
\usage{
get_senspe(roc)
}
\arguments{
\item{roc}{roc statistics obtain from et_roc function}
}
\value{
a tibble with two columns (sensitivities and sensibilities)
}
\description{
Function to get sensitivities ad specificities from a ROC statistics
}
\examples{
#find an example
}
