% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{train_mod}
\alias{train_mod}
\title{Function to fit a model}
\usage{
train_mod(xdata, yvec)
}
\arguments{
\item{xdata}{a dataframe with the predictors}

\item{yvec}{a vector with response variable values}
}
\value{
the results of the fitted model with function caret::train
}
\description{
Function to fit a model
}
\examples{
#find an example
}
